Implement IDisposable interface in Application class and update asset handling in AssetCache, AssetDatabase, and AssetRepository to use AssetGuid directly instead of its Value property. Adjust GameLoop to ensure fixedTimeStep is cast to float for consistency.

+++ Engine/Application/GameLoop.cs
@@ -44,7 +44,7 @@ public class GameLoop
             while (_accumulator >= fixedTimeStep)
             {
                 FixedUpdate((float)fixedTimeStep);
-                _accumulator -= fixedTimeStep;
+                _accumulator -= (float)fixedTimeStep;
             }
 
             // Variable timestep update
